On branch main Changes to be committed:
modified:   homeworks/zippass_pthread/src/Controller.c
    Find passwords function is now in a new function called
    execute_p, which will handle all the necessary operations
    for the job to be divided between all threads.
modified:   homeworks/zippass_pthread/src/Controller.h
    added execute_p, not documented atm.

Note: My idea is to calculate the total ammount of combinations to be
made and divide this by the number of threads to evenly distribute
the workload. The number of threads will be calculated in the execute_p
function.