The OMP solution works.
modified:   homeworks/zippass_omp_mpi/src/password_crack.c

The error was related with a variable which was shared, but marked as
private. The function now generates the lengths in a serial way, but
when it has to iterate through the total combinations possible for that
length the threads are created to make it parallel.